Generator of Manifest File with Barcode and Google Support
This is a hybrid application build in Cordova 6.1.1 for Android 5.1.1
platform with list of clients and their details which include
container(s) + item(s) based on barcodes scanned with the camera and
stored data in the HTML5 local storage.
It allows to create pdf file with all scanned container(s) + item(s),
total quantities of each as well as client details such as Name,
Address, Email, Phone and Comment which will be send via email.
It also uses Google Maps to locate the client on the map and allow the
user to find the route via navigation.